move replace-task to new module
